Made the UI toggleable with the backtick key
Renamed some lifetime parameters to make them more readable.

Replaced all occurrences of "draw" in identifiers to "render" to be more
consistent.

Collected all the arguments of EguiRenderer::render_ui into the struct
UiRenderingDescriptor.

Changed the type of UiRenderingDescriptor::ui_renderers from using a Box
to using a reference so that callers will not be forced to box their
variables.

Added the module ui_state that will store states that need to persist
through frames.