Overhauled the app architecture to use bevy_ecs
Renamed UiRenderer to EguiRenderer, ui_renderers.rs to egui_renderer.rs,
DefaultUiRenderer to DefaultEguiRenderer, and EguiRenderer to
EguiRendererResource. Removed ui_renderer.rs because it was no longer
needed. There used to be this "ui" and "egui" confusion because the
low-level egui renderer was already named EguiRenderer, so the actual
high-level UI renderer had to be named UiRenderer. Now that ECS is in
place, we have EguiRendererResource instead of EguiRenderer, so this
confusion no longer exists.