WIP: added some infrastructural types used for rendering
Added some basic types that will be used for the implementation of the
renderer. Right now there are a lot of dead_code warnings, but they will
go away later.

Added build.rs that will hard link the contents of the assets/ directory
next to the built executable, or create one if it doesn't exist, so that
assets_manager can access it.

Removed include_dir because shaders will now be loaded at runtime using
assets_manager instead. Also removed shaders.rs because shaders will now
be placed under the assets/ directory.

Added the diagrams/ directory to put some explanation diagrams for
concepts that are hard to explain in comments.

Moved World creation logic to its own module bevy_world.rs.

Renamed App::main_schedule to App::update_schedule to better show the
intent that it will be run every frame.

Added a log message that will print the Limits of the Adapter when it is
created, and used the reported Limits to create the Device so that it
can be used to its full potential, instead of only using the default
Limits.

Made main return Result<(), Box<dyn Error>> instead.

In src/bevy_resources/egui.rs, cited the source of the comment in
render_ui that I copied.